modellessdrf/requirements.txt
django==2.0
djangorestframework==3.7.3
httpie==0.9.9
django-cors-headers==2.1.0
gunicorn==19.7.1
